+++ libs/base_core/src/text/char/namespace.rs
@@ -7,7 +7,7 @@
 // - methods over u32
 // - methods over u8
 
-use crate::unwrap;
+use crate::{ASCII_TABLE, unwrap};
 
 #[doc = crate::_TAG_TEXT!()]
 #[doc = crate::_TAG_NAMESPACE!()]
@@ -16,20 +16,6 @@ use crate::unwrap;
 /// See also [`Str`][crate::Str], [`ExtMem`][crate::ExtMem],
 #[derive(Debug)]
 pub struct Char;
-
-/// Global ASCII lookup table for fast UTF-8 encoding.
-#[rustfmt::skip]
-pub(crate) static ASCII_TABLE: [&str; 128] = [
-    "\0", "\x01", "\x02", "\x03", "\x04", "\x05", "\x06", "\x07", "\x08", "\t", "\n",
-    "\x0B", "\x0C", "\r", "\x0E", "\x0F", "\x10", "\x11", "\x12", "\x13", "\x14", "\x15",
-    "\x16", "\x17", "\x18", "\x19", "\x1A", "\x1B", "\x1C", "\x1D", "\x1E", "\x1F",
-    " ", "!", "\"", "#", "$", "%", "&", "'", "(", ")", "*", "+", ",", "-", ".", "/",
-    "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", ":", ";", "<", "=", ">", "?",
-    "@", "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O",
-    "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "[", "\\", "]", "^", "_",
-    "`", "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o",
-    "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "{", "|", "}", "~", "\x7F",
-];
 
 /// # Methods over `char`
 #[rustfmt::skip]